Correction to ECMAScript import statement
As the file /src/index.js exports both Editor and Viewer
the import given in the readme is incorrect.

```import Editor from '@toast-ui/react-editor'```

imports an object called Editor with props of Editor.Editor
and Editor.Viewer.

This throws the following errors:

```
React.createElement: type is invalid -- expected a string (for built-in components)
        or a class/function (for composite components) but got: undefined.
        You likely forgot to export your component from the file it's defined in,
        or you might have mixed up default and named imports.
```